Py3 (#159)
* py3 fixes

* fibermap now Table not numpy array

* update .travis.yml and changes.rst prior to py3 PR

* update .travis.yml for py3.5 tests

* astropy 1.2.1 for tests

* let travis astropy 1.2.1 pick numpy version

* use default scipy version

* travis tests with specsim v0.5

+++ py/desisim/templates.py
@@ -1179,7 +1179,7 @@ class SUPERSTAR(object):
                 # Shuffle the basis templates and then split them into ~equal chunks, so
                 # we can speed up the calculations below.
                 chunksize = np.min((nbase, 50))
-                nchunk = long(np.ceil(nbase / chunksize))
+                nchunk = int(np.ceil(nbase / chunksize))
 
                 alltemplateid = np.tile(np.arange(nbase), (nmodel, 1))
                 for tempid in alltemplateid:
